Ownership can be transferred without confirmation
Fixed issues -1

An incorrect use of the execute_update_config, update_config or change_owner functions in some contracts can set owner to an invalid address and inadvertently lose control of them, which cannot be undone in any way. The affected contracts are the following:

krp-oracle/oracle_pyth
It is recommended to split ownership transfer functionality into set_owner and accept_ownership functions. The latter function allows the transfer to be completed by the recipient.